requirements.txt
requests
flask
PyYAML
